requirements.txt
numpy
scipy
scikit-image
scikit-learn
pillow
SimpleITK
opencv-python
opencv-contrib-python
matplotlib